Refactor cell_inds to use uint32, not uint8
This is to support the new octree frontend for Enzo-E, which can have
enough cells per oct to overflow a uint8. Uint16s might be able to be
used instead if there are no datasets with nz > 40.

+++ yt/frontends/artio/_artio_caller.pyx
@@ -974,7 +974,7 @@ cdef class ARTIOOctreeContainer(SparseOctreeContainer):
     @cython.cdivision(True)
     def fill_sfc(self,
              np.ndarray[np.uint8_t, ndim=1] levels,
-             np.ndarray[np.uint8_t, ndim=1] cell_inds,
+             np.ndarray[np.uint32_t, ndim=1] cell_inds,
              np.ndarray[np.int64_t, ndim=1] file_inds,
              np.ndarray[np.int64_t, ndim=1] domain_counts,
              field_indices, dest_fields):
