+++ yt/geometry/oct_container.pyx
@@ -681,14 +681,14 @@ cdef class OctreeContainer:
                         num_cells = -1):
         # We create oct arrays of the correct size
         cdef np.ndarray[np.uint8_t, ndim=1] levels
-        cdef np.ndarray[np.uint8_t, ndim=1] cell_inds
+        cdef np.ndarray[np.uint32_t, ndim=1] cell_inds
         cdef np.ndarray[np.int64_t, ndim=1] file_inds
         if num_cells < 0:
             num_cells = selector.count_oct_cells(self, domain_id)
         # Initialize variables with dummy values
         levels = np.full(num_cells, 255, dtype="uint8")
         file_inds = np.full(num_cells, -1, dtype="int64")
-        cell_inds = np.full(num_cells, 8, dtype="uint8")
+        cell_inds = np.full(num_cells, 8, dtype="uint32")
         cdef oct_visitors.FillFileIndicesO visitor_o
         cdef oct_visitors.FillFileIndicesR visitor_r
         if self.fill_style == "r":
@@ -742,7 +742,7 @@ cdef class OctreeContainer:
     @cython.cdivision(True)
     def fill_level(self, int level,
                    np.ndarray[np.uint8_t, ndim=1] levels,
-                   np.ndarray[np.uint8_t, ndim=1] cell_inds,
+                   np.ndarray[np.uint32_t, ndim=1] cell_inds,
                    np.ndarray[np.int64_t, ndim=1] file_inds,
                    dest_fields, source_fields,
                    np.int64_t offset = 0):
@@ -793,7 +793,7 @@ cdef class OctreeContainer:
         -------
         oct_inds : int64 ndarray (nocts*8, )
             The on-domain index of the octs containing each cell
-        cell_inds : uint8 ndarray (nocts*8, )
+        cell_inds : uint32 ndarray (nocts*8, )
             The index of the cell in its parent oct
 
         Note
@@ -805,10 +805,10 @@ cdef class OctreeContainer:
 
         cdef NeighbourCellIndexVisitor visitor
 
-        cdef np.uint8_t[::1] cell_inds
+        cdef np.uint32_t[::1] cell_inds
         cdef np.int64_t[::1] oct_inds
 
-        cell_inds = np.full(num_octs*4**3, 8, dtype=np.uint8)
+        cell_inds = np.full(num_octs*4**3, 8, dtype=np.uint32)
         oct_inds = np.full(num_octs*4**3, -1, dtype=np.int64)
 
         visitor = NeighbourCellIndexVisitor(self, -1, n_ghost_zones)
@@ -825,7 +825,7 @@ cdef class OctreeContainer:
     def fill_level_with_domain(
                    self, int level,
                    np.uint8_t[:] levels,
-                   np.uint8_t[:] cell_inds,
+                   np.uint32_t[:] cell_inds,
                    np.int64_t[:] file_inds,
                    np.int32_t[:] domains,
                    dict dest_fields,
@@ -881,7 +881,7 @@ cdef class OctreeContainer:
         -------
         levels : uint8, shape (num_cells,)
             The level of each cell of the super oct
-        cell_inds : uint8, shape (num_cells, )
+        cell_inds : uint32, shape (num_cells, )
             The index of each cell of the super oct within its own oct
         file_inds : int64, shape (num_cells, )
             The on-file position of the cell. See notes below.
@@ -918,12 +918,12 @@ cdef class OctreeContainer:
         cdef NeighbourCellVisitor visitor
 
         cdef np.ndarray[np.uint8_t, ndim=1] levels
-        cdef np.ndarray[np.uint8_t, ndim=1] cell_inds
+        cdef np.ndarray[np.uint32_t, ndim=1] cell_inds
         cdef np.ndarray[np.int64_t, ndim=1] file_inds
         cdef np.ndarray[np.int32_t, ndim=1] domains
         levels = np.full(num_cells, 255, dtype="uint8")
         file_inds = np.full(num_cells, -1, dtype="int64")
-        cell_inds = np.full(num_cells, 8, dtype="uint8")
+        cell_inds = np.full(num_cells, 8, dtype="uint32")
         domains = np.full(num_cells, -1, dtype="int32")
 
         visitor = NeighbourCellVisitor(self, -1, n_ghost_zones)
